chore: rename `recommendationsApi` to `repoDataApi`
- addional refactoring: remove prop-types
