+++ features/info/types.ts
@@ -2,6 +2,7 @@ export type RepoInfoTypes = {
   id: number;
   name: string;
   description: string;
+  fullName: string;
   forks: number;
   watchers: number;
   stargazersCount: number;
