+++ features/repos/components/ErrorState/index.tsx
@@ -7,16 +7,15 @@
 import React from "react";
 import get from "lodash/get";
 import { compose } from "redux";
-import PropTypes from "prop-types";
 import { injectIntl, IntlShape } from "react-intl";
 import { T, CustomCard } from "@common";
-import { IResponse } from "@features/repos/api/getRecommendations";
+import { IResponse } from "@features/repos/api/getRepoData";
 
 interface ErrorStateProps {
   intl: IntlShape;
   loading: boolean;
   reposData: IResponse | undefined;
-  reposError: string | undefined;
+  reposError?: string;
 }
 
 const ErrorState: React.FC<ErrorStateProps> = ({ intl, reposData, reposError, loading }) => {
@@ -36,24 +35,6 @@ const ErrorState: React.FC<ErrorStateProps> = ({ intl, reposData, reposError, lo
       <T id={repoError} />
     </CustomCard>
   ) : null;
-};
-
-const types = {
-  reposData: PropTypes.shape({
-    totalCount: PropTypes.number.isRequired,
-    incompleteResults: PropTypes.bool.isRequired,
-    items: PropTypes.array.isRequired,
-  }),
-};
-
-const { reposData } = types;
-
-ErrorState.propTypes = {
-  // @ts-expect-error intl is of type IntlShape, which is not possible in PropTypes
-  intl: PropTypes.object.isRequired,
-  loading: PropTypes.bool.isRequired,
-  reposError: PropTypes.string.isRequired,
-  reposData,
 };
 
 export default compose(injectIntl)(ErrorState);
