+++ store/slices/index.ts
@@ -1,10 +1,10 @@
 import { combineReducers } from "@reduxjs/toolkit";
 import repos from "./repos";
-import { recommendationsApi } from "@features/repos/api/getRecommendations";
+import { repoDataApi } from "@features/repos/api/getRepoData";
 import { repoInfoApi } from "@features/info/api/getRepoInfo";
 
 export default combineReducers({
   repos,
-  [recommendationsApi.reducerPath]: recommendationsApi.reducer,
+  [repoDataApi.reducerPath]: repoDataApi.reducer,
   [repoInfoApi.reducerPath]: repoInfoApi.reducer,
 });
